CodeFactor: eliminate our C style casts
These have no place in C++ code!

Unfortunately there are some in third party bits which can't be fixed here
but should be raised upstream.

This should help with #1650.

+++ src/TMap.cpp
@@ -312,7 +312,7 @@ void TMap::connectExitStub(int roomId, int dirType)
             }
         } else {
             //to avoid rooms on same plane from stealing stubs
-            if ((int)pR->x != rx) {
+            if (pR->x != rx) {
                 continue;
             }
         }
